+++ packages/data-schema/__tests__/CustomOperations.test-d.ts
@@ -1,0 +1,191 @@
+import type {
+  Equal,
+  Expect,
+  ExtractModelMeta,
+  Prettify,
+} from '@aws-amplify/data-schema-types';
+import { ClientSchema, a } from '../index';
+
+describe('custom operations return types', () => {
+  describe('when .ref() a basic custom type', () => {
+    it('generates correct return type', () => {
+      const schema = a.schema({
+        MyType: a.customType({
+          string: a.string().required(),
+          enum: a.enum(['hello', 'bye']),
+        }),
+        aQuery: a
+          .query()
+          .function('someHandler')
+          .returns(a.ref('MyType').required()),
+      });
+
+      type ReturnType = Prettify<
+        ExtractModelMeta<
+          ClientSchema<typeof schema>
+        >['customOperations']['aQuery']['returnType']
+      >;
+
+      type Expected = {
+        enum?: 'hello' | 'bye' | null | undefined;
+        string: string;
+      };
+
+      type _ = Expect<Equal<ReturnType, Expected>>;
+    });
+  });
+
+  describe('when .ref() a nested custom types', () => {
+    it('generates correct return type', () => {
+      const schema = a.schema({
+        Location: a.customType({
+          long: a.float(),
+          lang: a.float(),
+        }),
+        MyType: a.customType({
+          string: a.string().required(),
+          enum: a.enum(['hello', 'bye']),
+          meta: a.customType({
+            description: a.string().required(),
+            location: a.ref('Location'),
+          }),
+        }),
+        aQuery: a
+          .query()
+          .function('someHandler')
+          .returns(a.ref('MyType').required()),
+      });
+
+      type ReturnType = Prettify<
+        ExtractModelMeta<
+          ClientSchema<typeof schema>
+        >['customOperations']['aQuery']['returnType']
+      >;
+
+      type Expected = {
+        enum?: 'hello' | 'bye' | null | undefined;
+        meta?:
+          | {
+              location?:
+                | {
+                    long?: number | null | undefined;
+                    lang?: number | null | undefined;
+                  }
+                | null
+                | undefined;
+              description: string;
+            }
+          | null
+          | undefined;
+        string: string;
+      };
+
+      type _ = Expect<Equal<ReturnType, Expected>>;
+    });
+  });
+
+  describe('when .ref() a basic model', () => {
+    it('generates correct return type', () => {
+      const schema = a.schema({
+        MyModel: a.model({
+          string: a.string().required(),
+          enum: a.enum(['hello', 'bye']),
+        }),
+        aQuery: a
+          .query()
+          .function('someHandler')
+          .returns(a.ref('MyModel').required()),
+      });
+
+      type ReturnType = Prettify<
+        ExtractModelMeta<
+          ClientSchema<typeof schema>
+        >['customOperations']['aQuery']['returnType']
+      >;
+
+      type Expected = {
+        string: string;
+        readonly id: string;
+        readonly createdAt: string;
+        readonly updatedAt: string;
+        enum?: 'hello' | 'bye' | null | undefined;
+      };
+
+      type _ = Expect<Equal<ReturnType, Expected>>;
+    });
+  });
+
+  describe('when .ref() a model that has nested custom type', () => {
+    it('generates correct return type', () => {
+      const schema = a.schema({
+        MyModel: a.model({
+          string: a.string().required(),
+          enum: a.enum(['hello', 'bye']),
+          meta: a.customType({
+            description: a.string().required(),
+            location: a.ref('Location'),
+          }),
+        }),
+        Location: a.customType({
+          long: a.float(),
+          lang: a.float(),
+        }),
+        aQuery: a
+          .query()
+          .function('someHandler')
+          .returns(a.ref('MyModel').required()),
+      });
+
+      type ReturnType = Prettify<
+        ExtractModelMeta<
+          ClientSchema<typeof schema>
+        >['customOperations']['aQuery']['returnType']
+      >;
+
+      type Expected = {
+        string: string;
+        readonly id: string;
+        readonly createdAt: string;
+        readonly updatedAt: string;
+        enum?: 'hello' | 'bye' | null | undefined;
+        meta?:
+          | {
+              location?:
+                | {
+                    long?: number | null | undefined;
+                    lang?: number | null | undefined;
+                  }
+                | null
+                | undefined;
+              description: string;
+            }
+          | null
+          | undefined;
+      };
+
+      type _ = Expect<Equal<ReturnType, Expected>>;
+    });
+  });
+
+  describe('when.ref() a enum', () => {
+    it('generates correct return type', () => {
+      const schema = a.schema({
+        Value: a.enum(['succeeded', 'failed']),
+        aQuery: a
+          .query()
+          .function('someHandler')
+          .returns(a.ref('Value').required()),
+      });
+
+      type ReturnType = Prettify<
+        ExtractModelMeta<
+          ClientSchema<typeof schema>
+        >['customOperations']['aQuery']['returnType']
+      >;
+
+      type Expected = 'succeeded' | 'failed';
+
+      type _ = Expect<Equal<ReturnType, Expected>>;
+    });
+  });
+});