+++ packages/data-schema/__tests__/MappedTypes/CustomOperations.test-d.ts
@@ -2,9 +2,6 @@ import type { Equal, Expect } from '@aws-amplify/data-schema-types';
 import { a } from '../../index';
 import { RefType } from '../../src/RefType';
 import {
-  Bag,
-  CustomOpReturnType,
-  ResolveRef,
   CustomOpShapes,
   CustomOpArguments,
 } from '../../src/MappedTypes/CustomOperations';
@@ -25,60 +22,6 @@ describe('Custom Operations mapper utils', () => {
       C: 'whatever c';
     };
   };
-
-  type SimpleBag = Bag<Schema, NonModelTypes>;
-
-  test('can create a type bag from a schema and non-model types', () => {
-    type Expected = Schema & {
-      B: 'whatever b';
-    } & {
-      C: 'whatever c';
-    };
-
-    type T = Expect<Equal<SimpleBag, Expected>>;
-  });
-
-  test('can resolve model refs shape from a bag', () => {
-    const ref = a.ref('A');
-    type Actual = ResolveRef<Deref<typeof ref>, SimpleBag>;
-    type T = Expect<Equal<Actual, 'whatever a' | null>>;
-  });
-
-  test('can resolve custom type refs shape from a bag', () => {
-    const ref = a.ref('B');
-    type Actual = ResolveRef<Deref<typeof ref>, SimpleBag>;
-    type T = Expect<Equal<Actual, 'whatever b' | null>>;
-  });
-
-  test('can resolve enum refs shape from a bag', () => {
-    const ref = a.ref('C');
-    type Actual = ResolveRef<Deref<typeof ref>, SimpleBag>;
-    type T = Expect<Equal<Actual, 'whatever c' | null>>;
-  });
-
-  test('can resolve scalar return type', () => {
-    const op = a
-      .query()
-      .arguments({
-        inputParam: a.string(),
-      })
-      .returns(a.string());
-
-    type Actual = CustomOpReturnType<OpShape<typeof op>, { a: 'unused' }>;
-    type T = Expect<Equal<Actual, string | null>>;
-  });
-
-  test('can resolve a ref from a bag', () => {
-    const op = a
-      .query()
-      .arguments({
-        inputParam: a.string(),
-      })
-      .returns(a.ref('a'));
-
-    type Actual = CustomOpReturnType<OpShape<typeof op>, { a: 'sentinel' }>;
-    type T = Expect<Equal<Actual, 'sentinel' | null>>;
-  });
 
   test('can resolve arguments from custom op', () => {
     const aMutation = a