+++ packages/data-schema/src/SchemaProcessor.ts
@@ -14,11 +14,18 @@ import {
 } from './ModelRelationalField';
 import type { ModelType, InternalModel } from './ModelType';
 import type { InternalModelIndexType } from './ModelIndex';
-import { Authorization, accessData } from './Authorization';
+import {
+  type Authorization,
+  type ResourceAuthorization,
+  type SchemaAuthorization,
+  accessData,
+  accessSchemaData,
+} from './Authorization';
 import {
   DerivedApiDefinition,
   JsResolver,
   JsResolverEntry,
+  FunctionSchemaAccess,
 } from '@aws-amplify/data-schema-types';
 import type { InternalRef, RefType } from './RefType';
 import type { EnumType } from './EnumType';
@@ -223,21 +230,21 @@ function modelFieldToGql(fieldDef: ModelFieldDef) {
 }
 
 function refFieldToGql(fieldDef: RefFieldDef): string {
-  const { link, required } = fieldDef;
+  const { link, valueRequired, array, arrayRequired } = fieldDef;
 
   let field = link;
 
-  if (required === true) {
+  if (valueRequired === true) {
     field += '!';
   }
 
-  // if (array) {
-  //   field = `[${field}]`;
-  // }
+  if (array === true) {
+    field = `[${field}]`;
+  }
 
-  // if (arrayRequired === true) {
-  //   field += '!';
-  // }
+  if (arrayRequired === true) {
+    field += '!';
+  }
 
   return field;
 }
@@ -290,9 +297,9 @@ function customOperationToGql(
   if (functionRef) {
     gqlHandlerContent = `@function(name: "${functionRef}") `;
   } else if (databaseType === 'sql' && handler && brand === 'inlineSql') {
-    gqlHandlerContent = `@sql(statement: "${escapeGraphQlString(
+    gqlHandlerContent = `@sql(statement: ${escapeGraphQlString(
       String(getHandlerData(handler)),
-    )}") `;
+    )}) `;
   } else if (databaseType === 'sql' && handler && brand === 'sqlReference') {
     gqlHandlerContent = `@sql(reference: "${getHandlerData(handler)}") `;
   }
@@ -307,7 +314,7 @@ function customOperationToGql(
  * @returns The string with special charactars escaped
  */
 function escapeGraphQlString(str: string) {
-  return str.replace(/"/g, '\\"');
+  return JSON.stringify(str);
 }
 
 /**
@@ -377,6 +384,21 @@ function mergeFieldObjects(
     if (fields) addFields(result, fields);
   }
   return result;
+}
+
+/**
+ * Throws if resource/lambda auth is configured at the model or field level
+ *
+ * @param authorization A list of authorization rules.
+ */
+function validateAuth(authorization: Authorization<any, any, any>[] = []) {
+  for (const entry of authorization) {
+    if (ruleIsResourceAuth(entry)) {
+      throw new Error(
+        'Lambda resource authorization is only confiugrable at the schema level',
+      );
+    }
+  }
 }
 
 /**
@@ -743,9 +765,10 @@ function processFieldLevelAuthRules(
   } = {};
 
   for (const [fieldName, fieldDef] of Object.entries(fields)) {
-    const { authString, authFields: fieldAuthField } = calculateAuth(
-      (fieldDef as InternalField)?.data?.authorization || [],
-    );
+    const fieldAuth = (fieldDef as InternalField)?.data?.authorization || [];
+
+    validateAuth(fieldAuth);
+    const { authString, authFields: fieldAuthField } = calculateAuth(fieldAuth);
 
     if (authString) fieldLevelAuthRules[fieldName] = authString;
     if (fieldAuthField) {
@@ -913,9 +936,56 @@ const transformedSecondaryIndexesForModel = (
 
 type DatabaseType = 'dynamodb' | 'sql';
 
+const ruleIsResourceAuth = (
+  authRule: SchemaAuthorization<any, any, any>,
+): authRule is ResourceAuthorization => {
+  const data = accessSchemaData(authRule);
+  return data.strategy === 'resource';
+};
+
+/**
+ * Separates out lambda resource auth rules from remaining schema rules.
+ *
+ * @param authRules schema auth rules
+ */
+const extractFunctionSchemaAccess = (
+  authRules: SchemaAuthorization<any, any, any>[],
+): {
+  schemaAuth: Authorization<any, any, any>[];
+  functionSchemaAccess: FunctionSchemaAccess[];
+} => {
+  const schemaAuth: Authorization<any, any, any>[] = [];
+  const functionSchemaAccess: FunctionSchemaAccess[] = [];
+  const defaultActions: ['query', 'mutate', 'listen'] = [
+    'query',
+    'mutate',
+    'listen',
+  ];
+
+  for (const rule of authRules) {
+    if (ruleIsResourceAuth(rule)) {
+      const ruleData = accessSchemaData(rule);
+
+      const fnAccess = {
+        resourceProvider: ruleData.resource,
+        actions: ruleData.operations || defaultActions,
+      };
+      functionSchemaAccess.push(fnAccess);
+    } else {
+      schemaAuth.push(rule);
+    }
+  }
+
+  return { schemaAuth, functionSchemaAccess };
+};
+
 const schemaPreprocessor = (
   schema: InternalSchema,
-): { schema: string; jsFunctions: any[] } => {
+): {
+  schema: string;
+  jsFunctions: JsResolver[];
+  functionSchemaAccess: FunctionSchemaAccess[];
+} => {
   const gqlModels: string[] = [];
 
   const customQueries = [];
@@ -932,11 +1002,16 @@ const schemaPreprocessor = (
   const fkFields = allImpliedFKs(schema);
   const topLevelTypes = Object.entries(schema.data.types);
 
+  const { schemaAuth, functionSchemaAccess } = extractFunctionSchemaAccess(
+    schema.data.authorization,
+  );
+
   for (const [typeName, typeDef] of topLevelTypes) {
+    validateAuth(typeDef.data?.authorization);
     const mostRelevantAuthRules: Authorization<any, any, any>[] =
       typeDef.data?.authorization?.length > 0
         ? typeDef.data.authorization
-        : schema.data.authorization;
+        : schemaAuth;
 
     if (!isInternalModel(typeDef)) {
       if (isEnumType(typeDef)) {
@@ -1080,7 +1155,7 @@ const schemaPreprocessor = (
 
   const processedSchema = gqlModels.join('\n\n');
 
-  return { schema: processedSchema, jsFunctions };
+  return { schema: processedSchema, jsFunctions, functionSchemaAccess };
 };
 
 function validateCustomOperations(
@@ -1271,7 +1346,9 @@ function generateCustomOperationTypes({
 export function processSchema(arg: {
   schema: InternalSchema;
 }): DerivedApiDefinition {
-  const { schema, jsFunctions } = schemaPreprocessor(arg.schema);
+  const { schema, jsFunctions, functionSchemaAccess } = schemaPreprocessor(
+    arg.schema,
+  );
 
-  return { schema, functionSlots: [], jsFunctions };
+  return { schema, functionSlots: [], jsFunctions, functionSchemaAccess };
 }